+++ block/manager.go
@@ -1,23 +1,15 @@
 package block
 
 import (
-	"bytes"
 	"context"
 	"encoding/binary"
-	"encoding/hex"
-	"errors"
-	"fmt"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	"cosmossdk.io/log"
 	goheaderstore "github.com/celestiaorg/go-header/store"
-	ds "github.com/ipfs/go-datastore"
 	"github.com/libp2p/go-libp2p/core/crypto"
-	"google.golang.org/protobuf/proto"
-
-	"github.com/rollkit/go-sequencing"
 
 	coreda "github.com/rollkit/rollkit/core/da"
 	coreexecutor "github.com/rollkit/rollkit/core/execution"
@@ -28,46 +20,6 @@ import (
 	"github.com/rollkit/rollkit/pkg/queue"
 	"github.com/rollkit/rollkit/pkg/store"
 	"github.com/rollkit/rollkit/types"
-	pb "github.com/rollkit/rollkit/types/pb/rollkit/v1"
-)
-
-const (
-	// defaultLazySleepPercent is the percentage of block time to wait to accumulate transactions
-	// in lazy mode.
-	// A value of 10 for e.g. corresponds to 10% of the block time. Must be between 0 and 100.
-	defaultLazySleepPercent = 10
-
-	// defaultDABlockTime is used only if DABlockTime is not configured for manager
-	defaultDABlockTime = 15 * time.Second
-
-	// defaultBlockTime is used only if BlockTime is not configured for manager
-	defaultBlockTime = 1 * time.Second
-
-	// defaultLazyBlockTime is used only if LazyBlockTime is not configured for manager
-	defaultLazyBlockTime = 60 * time.Second
-
-	// defaultMempoolTTL is the number of blocks until transaction is dropped from mempool
-	defaultMempoolTTL = 25
-
-	// blockProtocolOverhead is the protocol overhead when marshaling the block to blob
-	// see: https://gist.github.com/tuxcanfly/80892dde9cdbe89bfb57a6cb3c27bae2
-	blockProtocolOverhead = 1 << 16
-
-	// maxSubmitAttempts defines how many times Rollkit will re-try to publish block to DA layer.
-	// This is temporary solution. It will be removed in future versions.
-	maxSubmitAttempts = 30
-
-	// Applies to most channels, 100 is a large enough buffer to avoid blocking
-	channelLength = 100
-
-	// Applies to the headerInCh, 10000 is a large enough number for headers per DA block.
-	headerInChLength = 10000
-
-	// DAIncludedHeightKey is the key used for persisting the da included height in store.
-	DAIncludedHeightKey = "d"
-
-	// LastBatchDataKey is the key used for persisting the last batch data in store.
-	LastBatchDataKey = "l"
 )
 
 var (
@@ -77,25 +29,6 @@ var (
 	// initialBackoff defines initial value for block submission backoff
 	initialBackoff = 100 * time.Millisecond
 )
-
-// NewHeaderEvent is used to pass header and DA height to headerInCh
-type NewHeaderEvent struct {
-	Header   *types.SignedHeader
-	DAHeight uint64
-}
-
-// NewDataEvent is used to pass header and DA height to headerInCh
-type NewDataEvent struct {
-	Data     *types.Data
-	DAHeight uint64
-}
-
-// BatchData is used to pass batch, time and data (da.IDs) to BatchQueue
-type BatchData struct {
-	*coresequencer.Batch
-	time.Time
-	Data [][]byte
-}
 
 // Manager is responsible for aggregating transactions into blocks.
 type Manager struct {
@@ -158,64 +91,6 @@ type Manager struct {
 	sequencer     coresequencer.Sequencer
 	lastBatchData [][]byte
 	bq            *queue.Queue[BatchData]
-}
-
-// getInitialState tries to load lastState from Store, and if it's not available it reads genesis.
-func getInitialState(ctx context.Context, genesis genesis.Genesis, store store.Store, exec coreexecutor.Executor, logger log.Logger) (types.State, error) {
-	// Load the state from store.
-	s, err := store.GetState(ctx)
-
-	if errors.Is(err, ds.ErrNotFound) {
-		logger.Info("No state found in store, initializing new state")
-
-		// Initialize genesis block explicitly
-		err = store.SaveBlockData(ctx,
-			&types.SignedHeader{Header: types.Header{
-				DataHash:        new(types.Data).Hash(),
-				ProposerAddress: genesis.ProposerAddress(),
-				BaseHeader: types.BaseHeader{
-					ChainID: genesis.ChainID,
-					Height:  genesis.InitialHeight,
-					Time:    uint64(genesis.GenesisDAStartHeight.UnixNano()),
-				}}},
-			&types.Data{},
-			&types.Signature{},
-		)
-		if err != nil {
-			return types.State{}, fmt.Errorf("failed to save genesis block: %w", err)
-		}
-
-		// If the user is starting a fresh chain (or hard-forking), we assume the stored state is empty.
-		// TODO(tzdybal): handle max bytes
-		stateRoot, _, err := exec.InitChain(ctx, genesis.GenesisDAStartHeight, genesis.InitialHeight, genesis.ChainID)
-		if err != nil {
-			logger.Error("error while initializing chain", "error", err)
-			return types.State{}, err
-		}
-
-		s := types.State{
-			Version:         types.Version{},
-			ChainID:         genesis.ChainID,
-			InitialHeight:   genesis.InitialHeight,
-			LastBlockHeight: genesis.InitialHeight - 1,
-			LastBlockTime:   genesis.GenesisDAStartHeight,
-			AppHash:         stateRoot,
-			DAHeight:        0,
-		}
-		return s, nil
-	} else if err != nil {
-		logger.Error("error while getting state", "error", err)
-		return types.State{}, err
-	} else {
-		// Perform a sanity-check to stop the user from
-		// using a higher genesis than the last stored state.
-		// if they meant to hard-fork, they should have cleared the stored State
-		if uint64(genesis.InitialHeight) > s.LastBlockHeight { //nolint:unconvert
-			return types.State{}, fmt.Errorf("genesis.InitialHeight (%d) is greater than last stored state's LastBlockHeight (%d)", genesis.InitialHeight, s.LastBlockHeight)
-		}
-	}
-
-	return s, nil
 }
 
 // NewManager creates new block Manager.
@@ -368,27 +243,6 @@ func (m *Manager) init(ctx context.Context) {
 	}
 }
 
-func (m *Manager) setDAIncludedHeight(ctx context.Context, newHeight uint64) error {
-	for {
-		currentHeight := m.daIncludedHeight.Load()
-		if newHeight <= currentHeight {
-			break
-		}
-		if m.daIncludedHeight.CompareAndSwap(currentHeight, newHeight) {
-			heightBytes := make([]byte, 8)
-			binary.BigEndian.PutUint64(heightBytes, newHeight)
-			return m.store.SetMetadata(ctx, DAIncludedHeightKey, heightBytes)
-		}
-	}
-	return nil
-}
-
-// GetDAIncludedHeight returns the rollup height at which all blocks have been
-// included in the DA
-func (m *Manager) GetDAIncludedHeight() uint64 {
-	return m.daIncludedHeight.Load()
-}
-
 // SetDALC is used to set DataAvailabilityLayerClient used by Manager.
 func (m *Manager) SetDALC(dalc coreda.Client) {
 	m.dalc = dalc
@@ -397,13 +251,6 @@ func (m *Manager) SetDALC(dalc coreda.Client) {
 // isProposer returns whether or not the manager is a proposer
 func isProposer(_ crypto.PrivKey, _ types.State) (bool, error) {
 	return true, nil
-}
-
-// SetLastState is used to set lastState used by Manager.
-func (m *Manager) SetLastState(state types.State) {
-	m.lastStateMtx.Lock()
-	defer m.lastStateMtx.Unlock()
-	m.lastState = state
 }
 
 // GetStoreHeight returns the manager's store height
@@ -431,185 +278,12 @@ func (m *Manager) IsDAIncluded(hash types.Hash) bool {
 	return m.headerCache.IsDAIncluded(hash.String())
 }
 
-// getRemainingSleep calculates the remaining sleep time based on config and a start time.
-func (m *Manager) getRemainingSleep(start time.Time) time.Duration {
-	elapsed := time.Since(start)
-	interval := m.config.Node.BlockTime.Duration
-
-	if m.config.Node.LazyAggregator {
-		if m.buildingBlock && elapsed >= interval {
-			// Special case to give time for transactions to accumulate if we
-			// are coming out of a period of inactivity.
-			return (interval * time.Duration(defaultLazySleepPercent) / 100)
-		} else if !m.buildingBlock {
-			interval = m.config.Node.LazyBlockTime.Duration
-		}
-	}
-
-	if elapsed < interval {
-		return interval - elapsed
-	}
-
-	return 0
-}
-
 // GetExecutor returns the executor used by the manager.
 //
 // Note: this is a temporary method to allow testing the manager.
 // It will be removed once the manager is fully integrated with the execution client.
 func (m *Manager) GetExecutor() coreexecutor.Executor {
 	return m.exec
-}
-
-// BatchRetrieveLoop is responsible for retrieving batches from the sequencer.
-func (m *Manager) BatchRetrieveLoop(ctx context.Context) {
-	m.logger.Info("Starting BatchRetrieveLoop")
-	batchTimer := time.NewTimer(0)
-	defer batchTimer.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-batchTimer.C:
-			start := time.Now()
-			m.logger.Debug("Attempting to retrieve next batch",
-				"chainID", m.genesis.ChainID,
-				"lastBatchData", m.lastBatchData)
-
-			req := coresequencer.GetNextBatchRequest{
-				RollupId:      []byte(m.genesis.ChainID),
-				LastBatchData: m.lastBatchData,
-			}
-
-			res, err := m.sequencer.GetNextBatch(ctx, req)
-			if err != nil {
-				m.logger.Error("error while retrieving batch", "error", err)
-				// Always reset timer on error
-				batchTimer.Reset(m.config.Node.BlockTime.Duration)
-				continue
-			}
-
-			if res != nil && res.Batch != nil {
-				m.logger.Debug("Retrieved batch",
-					"txCount", len(res.Batch.Transactions),
-					"timestamp", res.Timestamp)
-
-				m.bq.Add(BatchData{Batch: res.Batch, Time: res.Timestamp, Data: res.BatchData})
-				if len(res.Batch.Transactions) != 0 {
-					h := convertBatchDataToBytes(res.BatchData)
-					if err := m.store.SetMetadata(ctx, LastBatchDataKey, h); err != nil {
-						m.logger.Error("error while setting last batch hash", "error", err)
-					}
-					m.lastBatchData = res.BatchData
-				}
-
-			} else {
-				m.logger.Debug("No batch available")
-			}
-
-			// Always reset timer
-			elapsed := time.Since(start)
-			remainingSleep := m.config.Node.BlockTime.Duration - elapsed
-			if remainingSleep < 0 {
-				remainingSleep = 0
-			}
-			batchTimer.Reset(remainingSleep)
-		}
-	}
-}
-
-// AggregationLoop is responsible for aggregating transactions into rollup-blocks.
-func (m *Manager) AggregationLoop(ctx context.Context) {
-	initialHeight := m.genesis.InitialHeight //nolint:gosec
-	height := m.store.Height()
-	var delay time.Duration
-
-	// TODO(tzdybal): double-check when https://github.com/celestiaorg/rollmint/issues/699 is resolved
-	if height < initialHeight {
-		delay = time.Until(m.genesis.GenesisDAStartHeight.Add(m.config.Node.BlockTime.Duration))
-	} else {
-		lastBlockTime := m.getLastBlockTime()
-		delay = time.Until(lastBlockTime.Add(m.config.Node.BlockTime.Duration))
-	}
-
-	if delay > 0 {
-		m.logger.Info("Waiting to produce block", "delay", delay)
-		time.Sleep(delay)
-	}
-
-	// blockTimer is used to signal when to build a block based on the
-	// rollup block time. A timer is used so that the time to build a block
-	// can be taken into account.
-	blockTimer := time.NewTimer(0)
-	defer blockTimer.Stop()
-
-	// Lazy Aggregator mode.
-	// In Lazy Aggregator mode, blocks are built only when there are
-	// transactions or every LazyBlockTime.
-	if m.config.Node.LazyAggregator {
-		m.lazyAggregationLoop(ctx, blockTimer)
-		return
-	}
-
-	m.normalAggregationLoop(ctx, blockTimer)
-}
-
-func (m *Manager) lazyAggregationLoop(ctx context.Context, blockTimer *time.Timer) {
-	// start is used to track the start time of the block production period
-	start := time.Now()
-	// lazyTimer is used to signal when a block should be built in
-	// lazy mode to signal that the chain is still live during long
-	// periods of inactivity.
-	lazyTimer := time.NewTimer(0)
-	defer lazyTimer.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		// the m.bq.notifyCh channel is signalled when batch becomes available in the batch queue
-		case _, ok := <-m.bq.NotifyCh():
-			if ok && !m.buildingBlock {
-				// set the buildingBlock flag to prevent multiple calls to reset the time
-				m.buildingBlock = true
-				// Reset the block timer based on the block time.
-				blockTimer.Reset(m.getRemainingSleep(start))
-			}
-			continue
-		case <-lazyTimer.C:
-		case <-blockTimer.C:
-		}
-		// Define the start time for the block production period
-		start = time.Now()
-		if err := m.publishBlock(ctx); err != nil && ctx.Err() == nil {
-			m.logger.Error("error while publishing block", "error", err)
-		}
-		// unset the buildingBlocks flag
-		m.buildingBlock = false
-		// Reset the lazyTimer to produce a block even if there
-		// are no transactions as a way to signal that the chain
-		// is still live.
-		lazyTimer.Reset(m.getRemainingSleep(start))
-	}
-}
-
-func (m *Manager) normalAggregationLoop(ctx context.Context, blockTimer *time.Timer) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-blockTimer.C:
-			// Define the start time for the block production period
-			start := time.Now()
-			if err := m.publishBlock(ctx); err != nil && ctx.Err() == nil {
-				m.logger.Error("error while publishing block", "error", err)
-			}
-			// Reset the blockTimer to signal the next block production
-			// period based on the block time.
-			blockTimer.Reset(m.getRemainingSleep(start))
-		}
-	}
 }
 
 // HeaderSubmissionLoop is responsible for submitting blocks to the DA layer.
@@ -632,965 +306,10 @@ func (m *Manager) HeaderSubmissionLoop(ctx context.Context) {
 	}
 }
 
-func (m *Manager) handleEmptyDataHash(ctx context.Context, header *types.Header) {
-	headerHeight := header.Height()
-	if bytes.Equal(header.DataHash, dataHashForEmptyTxs) {
-		var lastDataHash types.Hash
-		var err error
-		var lastData *types.Data
-		if headerHeight > 1 {
-			_, lastData, err = m.store.GetBlockData(ctx, headerHeight-1)
-			if lastData != nil {
-				lastDataHash = lastData.Hash()
-			}
-		}
-		// if no error then populate data, otherwise just skip and wait for Data to be synced
-		if err == nil {
-			metadata := &types.Metadata{
-				ChainID:      header.ChainID(),
-				Height:       headerHeight,
-				Time:         header.BaseHeader.Time,
-				LastDataHash: lastDataHash,
-			}
-			d := &types.Data{
-				Metadata: metadata,
-			}
-			m.dataCache.SetItem(headerHeight, d)
-		}
-	}
-}
-
-// SyncLoop is responsible for syncing blocks.
-//
-// SyncLoop processes headers gossiped in P2P network to know what's the latest block height,
-// block data is retrieved from DA layer.
-func (m *Manager) SyncLoop(ctx context.Context) {
-	daTicker := time.NewTicker(m.config.DA.BlockTime.Duration)
-	defer daTicker.Stop()
-	blockTicker := time.NewTicker(m.config.Node.BlockTime.Duration)
-	defer blockTicker.Stop()
-	for {
-		select {
-		case <-daTicker.C:
-			m.sendNonBlockingSignalToRetrieveCh()
-		case <-blockTicker.C:
-			m.sendNonBlockingSignalToHeaderStoreCh()
-			m.sendNonBlockingSignalToDataStoreCh()
-		case headerEvent := <-m.headerInCh:
-			// Only validated headers are sent to headerInCh, so we can safely assume that headerEvent.header is valid
-			header := headerEvent.Header
-			daHeight := headerEvent.DAHeight
-			headerHash := header.Hash().String()
-			headerHeight := header.Height()
-			m.logger.Debug("header retrieved",
-				"height", headerHash,
-				"daHeight", daHeight,
-				"hash", headerHash,
-			)
-			if headerHeight <= m.store.Height() || m.headerCache.IsSeen(headerHash) {
-				m.logger.Debug("header already seen", "height", headerHeight, "block hash", headerHash)
-				continue
-			}
-			m.headerCache.SetItem(headerHeight, header)
-
-			m.sendNonBlockingSignalToHeaderStoreCh()
-			m.sendNonBlockingSignalToRetrieveCh()
-
-			// check if the dataHash is dataHashForEmptyTxs
-			// no need to wait for syncing Data, instead prepare now and set
-			// so that trySyncNextBlock can progress
-			m.handleEmptyDataHash(ctx, &header.Header)
-
-			err := m.trySyncNextBlock(ctx, daHeight)
-			if err != nil {
-				m.logger.Info("failed to sync next block", "error", err)
-				continue
-			}
-			m.headerCache.SetSeen(headerHash)
-		case dataEvent := <-m.dataInCh:
-			data := dataEvent.Data
-			daHeight := dataEvent.DAHeight
-			dataHash := data.Hash().String()
-			dataHeight := data.Metadata.Height
-			m.logger.Debug("data retrieved",
-				"height", dataHash,
-				"daHeight", daHeight,
-				"hash", dataHash,
-			)
-			if dataHeight <= m.store.Height() || m.dataCache.IsSeen(dataHash) {
-				m.logger.Debug("data already seen", "height", dataHeight, "data hash", dataHash)
-				continue
-			}
-			m.dataCache.SetItem(dataHeight, data)
-
-			m.sendNonBlockingSignalToDataStoreCh()
-
-			err := m.trySyncNextBlock(ctx, daHeight)
-			if err != nil {
-				m.logger.Info("failed to sync next block", "error", err)
-				continue
-			}
-			m.dataCache.SetSeen(dataHash)
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-func (m *Manager) sendNonBlockingSignalToHeaderStoreCh() {
-	select {
-	case m.headerStoreCh <- struct{}{}:
-	default:
-	}
-}
-
-func (m *Manager) sendNonBlockingSignalToDataStoreCh() {
-	select {
-	case m.dataStoreCh <- struct{}{}:
-	default:
-	}
-}
-
-func (m *Manager) sendNonBlockingSignalToRetrieveCh() {
-	select {
-	case m.retrieveCh <- struct{}{}:
-	default:
-	}
-}
-
-// trySyncNextBlock tries to execute as many blocks as possible from the blockCache.
-//
-//	Note: the blockCache contains only valid blocks that are not yet synced
-//
-// For every block, to be able to apply block at height h, we need to have its Commit. It is contained in block at height h+1.
-// If commit for block h+1 is available, we proceed with sync process, and remove synced block from sync cache.
-func (m *Manager) trySyncNextBlock(ctx context.Context, daHeight uint64) error {
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-		currentHeight := m.store.Height()
-		h := m.headerCache.GetItem(currentHeight + 1)
-		if h == nil {
-			m.logger.Debug("header not found in cache", "height", currentHeight+1)
-			return nil
-		}
-		d := m.dataCache.GetItem(currentHeight + 1)
-		if d == nil {
-			m.logger.Debug("data not found in cache", "height", currentHeight+1)
-			return nil
-		}
-
-		hHeight := h.Height()
-		m.logger.Info("Syncing header and data", "height", hHeight)
-		// Validate the received block before applying
-		if err := m.execValidate(m.lastState, h, d); err != nil {
-			return fmt.Errorf("failed to validate block: %w", err)
-		}
-		newState, err := m.applyBlock(ctx, h, d)
-		if err != nil {
-			if ctx.Err() != nil {
-				return err
-			}
-			// if call to applyBlock fails, we halt the node, see https://github.com/cometbft/cometbft/pull/496
-			panic(fmt.Errorf("failed to ApplyBlock: %w", err))
-		}
-		err = m.store.SaveBlockData(ctx, h, d, &h.Signature)
-		if err != nil {
-			return SaveBlockError{err}
-		}
-		_, err = m.execCommit(ctx, newState, h, d)
-		if err != nil {
-			return fmt.Errorf("failed to Commit: %w", err)
-		}
-
-		// Height gets updated
-		m.store.SetHeight(ctx, hHeight)
-
-		if daHeight > newState.DAHeight {
-			newState.DAHeight = daHeight
-		}
-		err = m.updateState(ctx, newState)
-		if err != nil {
-			m.logger.Error("failed to save updated state", "error", err)
-		}
-		m.headerCache.DeleteItem(currentHeight + 1)
-		m.dataCache.DeleteItem(currentHeight + 1)
-	}
-}
-
-// HeaderStoreRetrieveLoop is responsible for retrieving headers from the Header Store.
-func (m *Manager) HeaderStoreRetrieveLoop(ctx context.Context) {
-	lastHeaderStoreHeight := uint64(0)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.headerStoreCh:
-		}
-		headerStoreHeight := m.headerStore.Height()
-		if headerStoreHeight > lastHeaderStoreHeight {
-			headers, err := m.getHeadersFromHeaderStore(ctx, lastHeaderStoreHeight+1, headerStoreHeight)
-			if err != nil {
-				m.logger.Error("failed to get headers from Header Store", "lastHeaderHeight", lastHeaderStoreHeight, "headerStoreHeight", headerStoreHeight, "errors", err.Error())
-				continue
-			}
-			daHeight := atomic.LoadUint64(&m.daHeight)
-			for _, header := range headers {
-				// Check for shut down event prior to logging
-				// and sending header to headerInCh. The reason
-				// for checking for the shutdown event
-				// separately is due to the inconsistent nature
-				// of the select statement when multiple cases
-				// are satisfied.
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-				// early validation to reject junk headers
-				if !m.isUsingExpectedCentralizedSequencer(header) {
-					continue
-				}
-				m.logger.Debug("header retrieved from p2p header sync", "headerHeight", header.Height(), "daHeight", daHeight)
-				m.headerInCh <- NewHeaderEvent{header, daHeight}
-			}
-		}
-		lastHeaderStoreHeight = headerStoreHeight
-	}
-}
-
-// DataStoreRetrieveLoop is responsible for retrieving data from the Data Store.
-func (m *Manager) DataStoreRetrieveLoop(ctx context.Context) {
-	lastDataStoreHeight := uint64(0)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.dataStoreCh:
-		}
-		dataStoreHeight := m.dataStore.Height()
-		if dataStoreHeight > lastDataStoreHeight {
-			data, err := m.getDataFromDataStore(ctx, lastDataStoreHeight+1, dataStoreHeight)
-			if err != nil {
-				m.logger.Error("failed to get data from Data Store", "lastDataStoreHeight", lastDataStoreHeight, "dataStoreHeight", dataStoreHeight, "errors", err.Error())
-				continue
-			}
-			daHeight := atomic.LoadUint64(&m.daHeight)
-			for _, d := range data {
-				// Check for shut down event prior to logging
-				// and sending header to dataInCh. The reason
-				// for checking for the shutdown event
-				// separately is due to the inconsistent nature
-				// of the select statement when multiple cases
-				// are satisfied.
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-				//TODO: remove junk if possible
-				m.logger.Debug("data retrieved from p2p data sync", "dataHeight", d.Metadata.Height, "daHeight", daHeight)
-				m.dataInCh <- NewDataEvent{d, daHeight}
-			}
-		}
-		lastDataStoreHeight = dataStoreHeight
-	}
-}
-
-func (m *Manager) getHeadersFromHeaderStore(ctx context.Context, startHeight, endHeight uint64) ([]*types.SignedHeader, error) {
-	if startHeight > endHeight {
-		return nil, fmt.Errorf("startHeight (%d) is greater than endHeight (%d)", startHeight, endHeight)
-	}
-	headers := make([]*types.SignedHeader, endHeight-startHeight+1)
-	for i := startHeight; i <= endHeight; i++ {
-		header, err := m.headerStore.GetByHeight(ctx, i)
-		if err != nil {
-			return nil, err
-		}
-		headers[i-startHeight] = header
-	}
-	return headers, nil
-}
-
-func (m *Manager) getDataFromDataStore(ctx context.Context, startHeight, endHeight uint64) ([]*types.Data, error) {
-	if startHeight > endHeight {
-		return nil, fmt.Errorf("startHeight (%d) is greater than endHeight (%d)", startHeight, endHeight)
-	}
-	data := make([]*types.Data, endHeight-startHeight+1)
-	for i := startHeight; i <= endHeight; i++ {
-		d, err := m.dataStore.GetByHeight(ctx, i)
-		if err != nil {
-			return nil, err
-		}
-		data[i-startHeight] = d
-	}
-	return data, nil
-}
-
-// RetrieveLoop is responsible for interacting with DA layer.
-func (m *Manager) RetrieveLoop(ctx context.Context) {
-	// blockFoundCh is used to track when we successfully found a block so
-	// that we can continue to try and find blocks that are in the next DA height.
-	// This enables syncing faster than the DA block time.
-	headerFoundCh := make(chan struct{}, 1)
-	defer close(headerFoundCh)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-m.retrieveCh:
-		case <-headerFoundCh:
-		}
-		daHeight := atomic.LoadUint64(&m.daHeight)
-		err := m.processNextDAHeader(ctx)
-		if err != nil && ctx.Err() == nil {
-			// if the requested da height is not yet available, wait silently, otherwise log the error and wait
-			if !errors.Is(err, ErrHeightFromFutureStr) {
-				m.logger.Error("failed to retrieve block from DALC", "daHeight", daHeight, "errors", err.Error())
-			}
-			continue
-		}
-		// Signal the headerFoundCh to try and retrieve the next block
-		select {
-		case headerFoundCh <- struct{}{}:
-		default:
-		}
-		atomic.AddUint64(&m.daHeight, 1)
-	}
-}
-
-func (m *Manager) processNextDAHeader(ctx context.Context) error {
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	default:
-	}
-
-	// TODO(tzdybal): extract configuration option
-	maxRetries := 10
-	daHeight := atomic.LoadUint64(&m.daHeight)
-
-	var err error
-	m.logger.Debug("trying to retrieve block from DA", "daHeight", daHeight)
-	for r := 0; r < maxRetries; r++ {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-		headerResp, fetchErr := m.fetchHeaders(ctx, daHeight)
-		if fetchErr == nil {
-			if headerResp.Code == coreda.StatusNotFound {
-				m.logger.Debug("no header found", "daHeight", daHeight, "reason", headerResp.Message)
-				return nil
-			}
-			m.logger.Debug("retrieved potential headers", "n", len(headerResp.Data), "daHeight", daHeight)
-			for _, bz := range headerResp.Data {
-				header := new(types.SignedHeader)
-				// decode the header
-				var headerPb pb.SignedHeader
-				err := proto.Unmarshal(bz, &headerPb)
-				if err != nil {
-					m.logger.Error("failed to unmarshal header", "error", err)
-					continue
-				}
-				err = header.FromProto(&headerPb)
-				if err != nil {
-					m.logger.Error("failed to unmarshal header", "error", err)
-					continue
-				}
-				// early validation to reject junk headers
-				if !m.isUsingExpectedCentralizedSequencer(header) {
-					m.logger.Debug("skipping header from unexpected sequencer",
-						"headerHeight", header.Height(),
-						"headerHash", header.Hash().String())
-					continue
-				}
-				blockHash := header.Hash().String()
-				m.headerCache.SetDAIncluded(blockHash)
-				err = m.setDAIncludedHeight(ctx, header.Height())
-				if err != nil {
-					return err
-				}
-				m.logger.Info("block marked as DA included", "blockHeight", header.Height(), "blockHash", blockHash)
-				if !m.headerCache.IsSeen(blockHash) {
-					// Check for shut down event prior to logging
-					// and sending block to blockInCh. The reason
-					// for checking for the shutdown event
-					// separately is due to the inconsistent nature
-					// of the select statement when multiple cases
-					// are satisfied.
-					select {
-					case <-ctx.Done():
-						return fmt.Errorf("unable to send block to blockInCh, context done: %w", ctx.Err())
-					default:
-					}
-					m.headerInCh <- NewHeaderEvent{header, daHeight}
-				}
-			}
-			return nil
-		}
-
-		// Track the error
-		err = errors.Join(err, fetchErr)
-		// Delay before retrying
-		select {
-		case <-ctx.Done():
-			return err
-		case <-time.After(100 * time.Millisecond):
-		}
-	}
-	return err
-}
-
-func (m *Manager) isUsingExpectedCentralizedSequencer(header *types.SignedHeader) bool {
-	return bytes.Equal(header.ProposerAddress, m.genesis.ProposerAddress()) && header.ValidateBasic() == nil
-}
-
-func (m *Manager) fetchHeaders(ctx context.Context, daHeight uint64) (coreda.ResultRetrieve, error) {
-	var err error
-	ctx, cancel := context.WithTimeout(ctx, 60*time.Second) //TODO: make this configurable
-	defer cancel()
-	headerRes := m.dalc.Retrieve(ctx, daHeight)
-	if headerRes.Code == coreda.StatusError {
-		err = fmt.Errorf("failed to retrieve block: %s", headerRes.Message)
-	}
-	return headerRes, err
-}
-
 func (m *Manager) getSignature(header types.Header) (types.Signature, error) {
 	b, err := header.MarshalBinary()
 	if err != nil {
 		return nil, err
 	}
 	return m.proposerKey.Sign(b)
-}
-
-func (m *Manager) getTxsFromBatch() (*BatchData, error) {
-	batch := m.bq.Next()
-	if batch == nil {
-		// batch is nil when there is nothing to process
-		return nil, ErrNoBatch
-	}
-	return &BatchData{Batch: batch.Batch, Time: batch.Time, Data: batch.Data}, nil
-}
-
-func (m *Manager) publishBlock(ctx context.Context) error {
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	default:
-	}
-
-	if !m.isProposer {
-		return ErrNotProposer
-	}
-
-	if m.config.Node.MaxPendingBlocks != 0 && m.pendingHeaders.numPendingHeaders() >= m.config.Node.MaxPendingBlocks {
-		return fmt.Errorf("refusing to create block: pending blocks [%d] reached limit [%d]",
-			m.pendingHeaders.numPendingHeaders(), m.config.Node.MaxPendingBlocks)
-	}
-
-	var (
-		lastSignature  *types.Signature
-		lastHeaderHash types.Hash
-		lastDataHash   types.Hash
-		lastHeaderTime time.Time
-		err            error
-	)
-	height := m.store.Height()
-	newHeight := height + 1
-	// this is a special case, when first block is produced - there is no previous commit
-	if newHeight <= m.genesis.InitialHeight {
-		// Special handling for genesis block
-		lastSignature = &types.Signature{}
-	} else {
-		lastSignature, err = m.store.GetSignature(ctx, height)
-		if err != nil {
-			return fmt.Errorf("error while loading last commit: %w", err)
-		}
-		lastHeader, lastData, err := m.store.GetBlockData(ctx, height)
-		if err != nil {
-			return fmt.Errorf("error while loading last block: %w", err)
-		}
-		lastHeaderHash = lastHeader.Hash()
-		lastDataHash = lastData.Hash()
-		lastHeaderTime = lastHeader.Time()
-	}
-
-	var (
-		header    *types.SignedHeader
-		data      *types.Data
-		signature types.Signature
-	)
-
-	// Check if there's an already stored block at a newer height
-	// If there is use that instead of creating a new block
-	pendingHeader, pendingData, err := m.store.GetBlockData(ctx, newHeight)
-	if err == nil {
-		m.logger.Info("Using pending block", "height", newHeight)
-		header = pendingHeader
-		data = pendingData
-	} else {
-		execTxs, err := m.exec.GetTxs(ctx)
-		if err != nil {
-			m.logger.Error("failed to get txs from executor", "err", err)
-			// Continue but log the state
-			m.logger.Info("Current state",
-				"height", height,
-				"isProposer", m.isProposer,
-				"pendingHeaders", m.pendingHeaders.numPendingHeaders())
-		}
-
-		m.logger.Debug("Submitting transaction to sequencer",
-			"txCount", len(execTxs))
-		_, err = m.sequencer.SubmitRollupBatchTxs(ctx, coresequencer.SubmitRollupBatchTxsRequest{
-			RollupId: sequencing.RollupId(m.genesis.ChainID),
-			Batch:    &coresequencer.Batch{Transactions: execTxs},
-		})
-		if err != nil {
-			m.logger.Error("failed to submit rollup transaction to sequencer",
-				"err", err,
-				"chainID", m.genesis.ChainID)
-			// Add retry logic or proper error handling
-
-			m.logger.Debug("Successfully submitted transaction to sequencer")
-		}
-
-		batchData, err := m.getTxsFromBatch()
-		if errors.Is(err, ErrNoBatch) {
-			m.logger.Debug("No batch available, creating empty block")
-			// Create an empty block instead of returning
-			batchData = &BatchData{
-				Batch: &coresequencer.Batch{Transactions: [][]byte{}},
-				Time:  time.Now().Round(0).UTC(),
-				Data:  [][]byte{},
-			}
-		} else if err != nil {
-			return fmt.Errorf("failed to get transactions from batch: %w", err)
-		}
-		// sanity check timestamp for monotonically increasing
-		if batchData.Time.Before(lastHeaderTime) {
-			return fmt.Errorf("timestamp is not monotonically increasing: %s < %s", batchData.Time, m.getLastBlockTime())
-		}
-		m.logger.Info("Creating and publishing block", "height", newHeight)
-		header, data, err = m.createBlock(ctx, newHeight, lastSignature, lastHeaderHash, batchData)
-		if err != nil {
-			return err
-		}
-		m.logger.Debug("block info", "num_tx", len(data.Txs))
-
-		/*
-		   here we set the SignedHeader.DataHash, and SignedHeader.Signature as a hack
-		   to make the block pass ValidateBasic() when it gets called by applyBlock on line 681
-		   these values get overridden on lines 687-698 after we obtain the IntermediateStateRoots.
-		*/
-		header.DataHash = data.Hash()
-		//header.Validators = m.getLastStateValidators()
-		//header.ValidatorHash = header.Validators.Hash()
-
-		signature, err = m.getSignature(header.Header)
-		if err != nil {
-			return err
-		}
-
-		// set the signature to current block's signed header
-		header.Signature = signature
-		err = m.store.SaveBlockData(ctx, header, data, &signature)
-		if err != nil {
-			return SaveBlockError{err}
-		}
-	}
-
-	newState, err := m.applyBlock(ctx, header, data)
-	if err != nil {
-		if ctx.Err() != nil {
-			return err
-		}
-		// if call to applyBlock fails, we halt the node, see https://github.com/cometbft/cometbft/pull/496
-		panic(err)
-	}
-	// Before taking the hash, we need updated ISRs, hence after ApplyBlock
-	header.Header.DataHash = data.Hash()
-
-	signature, err = m.getSignature(header.Header)
-	if err != nil {
-		return err
-	}
-
-	// set the signature to current block's signed header
-	header.Signature = signature
-
-	// append metadata to Data before validating and saving
-	data.Metadata = &types.Metadata{
-		ChainID:      header.ChainID(),
-		Height:       header.Height(),
-		Time:         header.BaseHeader.Time,
-		LastDataHash: lastDataHash,
-	}
-	// Validate the created block before storing
-	if err := m.execValidate(m.lastState, header, data); err != nil {
-		return fmt.Errorf("failed to validate block: %w", err)
-	}
-
-	headerHeight := header.Height()
-
-	headerHash := header.Hash().String()
-	m.headerCache.SetSeen(headerHash)
-
-	// SaveBlock commits the DB tx
-	err = m.store.SaveBlockData(ctx, header, data, &signature)
-	if err != nil {
-		return SaveBlockError{err}
-	}
-
-	// Commit the new state and block which writes to disk on the proxy app
-	appHash, err := m.execCommit(ctx, newState, header, data)
-	if err != nil {
-		return err
-	}
-	// Update app hash in state
-	newState.AppHash = appHash
-
-	// Update the store height before submitting to the DA layer but after committing to the DB
-	m.store.SetHeight(ctx, headerHeight)
-
-	newState.DAHeight = atomic.LoadUint64(&m.daHeight)
-	// After this call m.lastState is the NEW state returned from ApplyBlock
-	// updateState also commits the DB tx
-	m.logger.Debug("updating state", "newState", newState)
-	err = m.updateState(ctx, newState)
-	if err != nil {
-		return err
-	}
-	m.recordMetrics(data)
-	// Check for shut down event prior to sending the header and block to
-	// their respective channels. The reason for checking for the shutdown
-	// event separately is due to the inconsistent nature of the select
-	// statement when multiple cases are satisfied.
-	select {
-	case <-ctx.Done():
-		return fmt.Errorf("unable to send header and block, context done: %w", ctx.Err())
-	default:
-	}
-
-	// Publish header to channel so that header exchange service can broadcast
-	m.HeaderCh <- header
-
-	// Publish block to channel so that block exchange service can broadcast
-	m.DataCh <- data
-
-	m.logger.Debug("successfully proposed header", "proposer", hex.EncodeToString(header.ProposerAddress), "height", headerHeight)
-
-	return nil
-}
-
-func (m *Manager) recordMetrics(data *types.Data) {
-	m.metrics.NumTxs.Set(float64(len(data.Txs)))
-	m.metrics.TotalTxs.Add(float64(len(data.Txs)))
-	m.metrics.BlockSizeBytes.Set(float64(data.Size()))
-	m.metrics.CommittedHeight.Set(float64(data.Metadata.Height))
-}
-func (m *Manager) submitHeadersToDA(ctx context.Context) error {
-	submittedAllHeaders := false
-	var backoff time.Duration
-	headersToSubmit, err := m.pendingHeaders.getPendingHeaders(ctx)
-	if len(headersToSubmit) == 0 {
-		// There are no pending headers; return because there's nothing to do, but:
-		// - it might be caused by error, then err != nil
-		// - all pending headers are processed, then err == nil
-		// whatever the reason, error information is propagated correctly to the caller
-		return err
-	}
-	if err != nil {
-		// There are some pending blocks but also an error. It's very unlikely case - probably some error while reading
-		// headers from the store.
-		// The error is logged and normal processing of pending blocks continues.
-		m.logger.Error("error while fetching blocks pending DA", "err", err)
-	}
-	numSubmittedHeaders := 0
-	attempt := 0
-	maxBlobSize, err := m.dalc.MaxBlobSize(ctx)
-	if err != nil {
-		return err
-	}
-	initialMaxBlobSize := maxBlobSize
-	gasPrice := m.gasPrice
-	initialGasPrice := gasPrice
-
-daSubmitRetryLoop:
-	for !submittedAllHeaders && attempt < maxSubmitAttempts {
-		select {
-		case <-ctx.Done():
-			break daSubmitRetryLoop
-		case <-time.After(backoff):
-		}
-
-		headersBz := make([][]byte, len(headersToSubmit))
-		for i, header := range headersToSubmit {
-			headerPb, err := header.ToProto()
-			if err != nil {
-				// do we drop the header from attempting to be submitted?
-				return fmt.Errorf("failed to transform header to proto: %w", err)
-			}
-			headersBz[i], err = proto.Marshal(headerPb)
-			if err != nil {
-				// do we drop the header from attempting to be submitted?
-				return fmt.Errorf("failed to marshal header: %w", err)
-			}
-		}
-
-		ctx, cancel := context.WithTimeout(ctx, 60*time.Second) //TODO: make this configurable
-		defer cancel()
-		res := m.dalc.Submit(ctx, headersBz, maxBlobSize, gasPrice)
-		switch res.Code {
-		case coreda.StatusSuccess:
-			m.logger.Info("successfully submitted Rollkit headers to DA layer", "gasPrice", gasPrice, "daHeight", res.Height, "headerCount", res.SubmittedCount)
-			if res.SubmittedCount == uint64(len(headersToSubmit)) {
-				submittedAllHeaders = true
-			}
-			submittedBlocks, notSubmittedBlocks := headersToSubmit[:res.SubmittedCount], headersToSubmit[res.SubmittedCount:]
-			numSubmittedHeaders += len(submittedBlocks)
-			for _, block := range submittedBlocks {
-				m.headerCache.SetDAIncluded(block.Hash().String())
-				err = m.setDAIncludedHeight(ctx, block.Height())
-				if err != nil {
-					return err
-				}
-			}
-			lastSubmittedHeight := uint64(0)
-			if l := len(submittedBlocks); l > 0 {
-				lastSubmittedHeight = submittedBlocks[l-1].Height()
-			}
-			m.pendingHeaders.setLastSubmittedHeight(ctx, lastSubmittedHeight)
-			headersToSubmit = notSubmittedBlocks
-			// reset submission options when successful
-			// scale back gasPrice gradually
-			backoff = 0
-			maxBlobSize = initialMaxBlobSize
-			if m.gasMultiplier > 0 && gasPrice != -1 {
-				gasPrice = gasPrice / m.gasMultiplier
-				if gasPrice < initialGasPrice {
-					gasPrice = initialGasPrice
-				}
-			}
-			m.logger.Debug("resetting DA layer submission options", "backoff", backoff, "gasPrice", gasPrice, "maxBlobSize", maxBlobSize)
-		case coreda.StatusNotIncludedInBlock, coreda.StatusAlreadyInMempool:
-			m.logger.Error("DA layer submission failed", "error", res.Message, "attempt", attempt)
-			backoff = m.config.DA.BlockTime.Duration * time.Duration(m.config.DA.MempoolTTL) //nolint:gosec
-			if m.gasMultiplier > 0 && gasPrice != -1 {
-				gasPrice = gasPrice * m.gasMultiplier
-			}
-			m.logger.Info("retrying DA layer submission with", "backoff", backoff, "gasPrice", gasPrice, "maxBlobSize", maxBlobSize)
-
-		case coreda.StatusTooBig:
-			maxBlobSize = maxBlobSize / 4
-			fallthrough
-		default:
-			m.logger.Error("DA layer submission failed", "error", res.Message, "attempt", attempt)
-			backoff = m.exponentialBackoff(backoff)
-		}
-
-		attempt += 1
-	}
-
-	if !submittedAllHeaders {
-		return fmt.Errorf(
-			"failed to submit all blocks to DA layer, submitted %d blocks (%d left) after %d attempts",
-			numSubmittedHeaders,
-			len(headersToSubmit),
-			attempt,
-		)
-	}
-	return nil
-}
-
-func (m *Manager) exponentialBackoff(backoff time.Duration) time.Duration {
-	backoff *= 2
-	if backoff == 0 {
-		backoff = initialBackoff
-	}
-	if backoff > m.config.DA.BlockTime.Duration {
-		backoff = m.config.DA.BlockTime.Duration
-	}
-	return backoff
-}
-
-// Updates the state stored in manager's store along the manager's lastState
-func (m *Manager) updateState(ctx context.Context, s types.State) error {
-	m.logger.Debug("updating state", "newState", s)
-	m.lastStateMtx.Lock()
-	defer m.lastStateMtx.Unlock()
-	err := m.store.UpdateState(ctx, s)
-	if err != nil {
-		return err
-	}
-	m.lastState = s
-	m.metrics.Height.Set(float64(s.LastBlockHeight))
-	return nil
-}
-
-func (m *Manager) getLastBlockTime() time.Time {
-	m.lastStateMtx.RLock()
-	defer m.lastStateMtx.RUnlock()
-	return m.lastState.LastBlockTime
-}
-
-func (m *Manager) createBlock(ctx context.Context, height uint64, lastSignature *types.Signature, lastHeaderHash types.Hash, batchData *BatchData) (*types.SignedHeader, *types.Data, error) {
-	m.lastStateMtx.RLock()
-	defer m.lastStateMtx.RUnlock()
-	return m.execCreateBlock(ctx, height, lastSignature, lastHeaderHash, m.lastState, batchData)
-}
-
-func (m *Manager) applyBlock(ctx context.Context, header *types.SignedHeader, data *types.Data) (types.State, error) {
-	m.lastStateMtx.RLock()
-	defer m.lastStateMtx.RUnlock()
-	return m.execApplyBlock(ctx, m.lastState, header, data)
-}
-
-func (m *Manager) execValidate(_ types.State, _ *types.SignedHeader, _ *types.Data) error {
-	// TODO(tzdybal): implement
-	return nil
-}
-
-func (m *Manager) execCommit(ctx context.Context, newState types.State, h *types.SignedHeader, _ *types.Data) ([]byte, error) {
-	err := m.exec.SetFinal(ctx, h.Height())
-	return newState.AppHash, err
-}
-
-func (m *Manager) execCreateBlock(_ context.Context, height uint64, lastSignature *types.Signature, lastHeaderHash types.Hash, lastState types.State, batchData *BatchData) (*types.SignedHeader, *types.Data, error) {
-	data := batchData.Data
-	batchdata := convertBatchDataToBytes(data)
-	header := &types.SignedHeader{
-		Header: types.Header{
-			Version: types.Version{
-				Block: lastState.Version.Block,
-				App:   lastState.Version.App,
-			},
-			BaseHeader: types.BaseHeader{
-				ChainID: lastState.ChainID,
-				Height:  height,
-				Time:    uint64(batchData.Time.UnixNano()), //nolint:gosec // why is time unix? (tac0turtle)
-			},
-			LastHeaderHash:  lastHeaderHash,
-			DataHash:        batchdata,
-			ConsensusHash:   make(types.Hash, 32),
-			AppHash:         lastState.AppHash,
-			ProposerAddress: m.genesis.ProposerAddress(),
-		},
-		Signature: *lastSignature,
-	}
-
-	blockData := &types.Data{
-		Txs: make(types.Txs, len(batchData.Batch.Transactions)),
-	}
-	for i := range batchData.Batch.Transactions {
-		blockData.Txs[i] = types.Tx(batchData.Batch.Transactions[i])
-	}
-
-	return header, blockData, nil
-}
-
-func (m *Manager) execApplyBlock(ctx context.Context, lastState types.State, header *types.SignedHeader, data *types.Data) (types.State, error) {
-	rawTxs := make([][]byte, len(data.Txs))
-	for i := range data.Txs {
-		rawTxs[i] = data.Txs[i]
-	}
-	newStateRoot, _, err := m.exec.ExecuteTxs(ctx, rawTxs, header.Height(), header.Time(), lastState.AppHash)
-	if err != nil {
-		return types.State{}, err
-	}
-
-	s, err := m.nextState(lastState, header, newStateRoot)
-	if err != nil {
-		return types.State{}, err
-	}
-
-	return s, nil
-}
-
-func (m *Manager) nextState(state types.State, header *types.SignedHeader, stateRoot []byte) (types.State, error) {
-	height := header.Height()
-
-	s := types.State{
-		Version:         state.Version,
-		ChainID:         state.ChainID,
-		InitialHeight:   state.InitialHeight,
-		LastBlockHeight: height,
-		LastBlockTime:   header.Time(),
-		AppHash:         stateRoot,
-	}
-	return s, nil
-}
-
-func convertBatchDataToBytes(batchData [][]byte) []byte {
-	// If batchData is nil or empty, return an empty byte slice
-	if len(batchData) == 0 {
-		return []byte{}
-	}
-
-	// For a single item, we still need to length-prefix it for consistency
-	// First, calculate the total size needed
-	// Format: 4 bytes (length) + data for each entry
-	totalSize := 0
-	for _, data := range batchData {
-		totalSize += 4 + len(data) // 4 bytes for length prefix + data length
-	}
-
-	// Allocate buffer with calculated capacity
-	result := make([]byte, 0, totalSize)
-
-	// Add length-prefixed data
-	for _, data := range batchData {
-		// Encode length as 4-byte big-endian integer
-		lengthBytes := make([]byte, 4)
-		binary.BigEndian.PutUint32(lengthBytes, uint32(len(data)))
-
-		// Append length prefix
-		result = append(result, lengthBytes...)
-
-		// Append actual data
-		result = append(result, data...)
-	}
-
-	return result
-}
-
-// bytesToBatchData converts a length-prefixed byte array back to a slice of byte slices
-func bytesToBatchData(data []byte) ([][]byte, error) {
-	if len(data) == 0 {
-		return [][]byte{}, nil
-	}
-
-	var result [][]byte
-	offset := 0
-
-	for offset < len(data) {
-		// Check if we have at least 4 bytes for the length prefix
-		if offset+4 > len(data) {
-			return nil, fmt.Errorf("corrupted data: insufficient bytes for length prefix at offset %d", offset)
-		}
-
-		// Read the length prefix
-		length := binary.BigEndian.Uint32(data[offset : offset+4])
-		offset += 4
-
-		// Check if we have enough bytes for the data
-		if offset+int(length) > len(data) {
-			return nil, fmt.Errorf("corrupted data: insufficient bytes for entry of length %d at offset %d", length, offset)
-		}
-
-		// Extract the data entry
-		entry := make([]byte, length)
-		copy(entry, data[offset:offset+int(length)])
-		result = append(result, entry)
-
-		// Move to the next entry
-		offset += int(length)
-	}
-
-	return result, nil
 }
