+++ block/state.go
@@ -1,0 +1,136 @@
+package block
+
+import (
+	"context"
+	"encoding/binary"
+	"errors"
+	"fmt"
+	"time"
+
+	"cosmossdk.io/log"
+	ds "github.com/ipfs/go-datastore"
+	coreexecutor "github.com/rollkit/rollkit/core/execution"
+	"github.com/rollkit/rollkit/pkg/genesis"
+	"github.com/rollkit/rollkit/pkg/store"
+	"github.com/rollkit/rollkit/types"
+)
+
+// getInitialState tries to load lastState from Store, and if it's not available it reads genesis.
+func getInitialState(ctx context.Context, genesis genesis.Genesis, store store.Store, exec coreexecutor.Executor, logger log.Logger) (types.State, error) {
+	// Load the state from store.
+	s, err := store.GetState(ctx)
+
+	if errors.Is(err, ds.ErrNotFound) {
+		logger.Info("No state found in store, initializing new state")
+
+		// Initialize genesis block explicitly
+		err = store.SaveBlockData(ctx,
+			&types.SignedHeader{Header: types.Header{
+				DataHash:        new(types.Data).Hash(),
+				ProposerAddress: genesis.ProposerAddress(),
+				BaseHeader: types.BaseHeader{
+					ChainID: genesis.ChainID,
+					Height:  genesis.InitialHeight,
+					Time:    uint64(genesis.GenesisDAStartHeight.UnixNano()),
+				}}},
+			&types.Data{},
+			&types.Signature{},
+		)
+		if err != nil {
+			return types.State{}, fmt.Errorf("failed to save genesis block: %w", err)
+		}
+
+		// If the user is starting a fresh chain (or hard-forking), we assume the stored state is empty.
+		// TODO(tzdybal): handle max bytes
+		stateRoot, _, err := exec.InitChain(ctx, genesis.GenesisDAStartHeight, genesis.InitialHeight, genesis.ChainID)
+		if err != nil {
+			logger.Error("error while initializing chain", "error", err)
+			return types.State{}, err
+		}
+
+		s := types.State{
+			Version:         types.Version{},
+			ChainID:         genesis.ChainID,
+			InitialHeight:   genesis.InitialHeight,
+			LastBlockHeight: genesis.InitialHeight - 1,
+			LastBlockTime:   genesis.GenesisDAStartHeight,
+			AppHash:         stateRoot,
+			DAHeight:        0,
+		}
+		return s, nil
+	} else if err != nil {
+		logger.Error("error while getting state", "error", err)
+		return types.State{}, err
+	} else {
+		// Perform a sanity-check to stop the user from
+		// using a higher genesis than the last stored state.
+		// if they meant to hard-fork, they should have cleared the stored State
+		if uint64(genesis.InitialHeight) > s.LastBlockHeight { //nolint:unconvert
+			return types.State{}, fmt.Errorf("genesis.InitialHeight (%d) is greater than last stored state's LastBlockHeight (%d)", genesis.InitialHeight, s.LastBlockHeight)
+		}
+	}
+
+	return s, nil
+}
+
+// Updates the state stored in manager's store along the manager's lastState
+func (m *Manager) updateState(ctx context.Context, s types.State) error {
+	m.logger.Debug("updating state", "newState", s)
+	m.lastStateMtx.Lock()
+	defer m.lastStateMtx.Unlock()
+	err := m.store.UpdateState(ctx, s)
+	if err != nil {
+		return err
+	}
+	m.lastState = s
+	m.metrics.Height.Set(float64(s.LastBlockHeight))
+	return nil
+}
+
+func (m *Manager) getLastBlockTime() time.Time {
+	m.lastStateMtx.RLock()
+	defer m.lastStateMtx.RUnlock()
+	return m.lastState.LastBlockTime
+}
+
+// SetLastState is used to set lastState used by Manager.
+func (m *Manager) SetLastState(state types.State) {
+	m.lastStateMtx.Lock()
+	defer m.lastStateMtx.Unlock()
+	m.lastState = state
+}
+
+func (m *Manager) nextState(state types.State, header *types.SignedHeader, stateRoot []byte) (types.State, error) {
+	height := header.Height()
+
+	s := types.State{
+		Version:         state.Version,
+		ChainID:         state.ChainID,
+		InitialHeight:   state.InitialHeight,
+		LastBlockHeight: height,
+		LastBlockTime:   header.Time(),
+		AppHash:         stateRoot,
+	}
+	return s, nil
+}
+
+func (m *Manager) setDAIncludedHeight(ctx context.Context, newHeight uint64) error {
+	for {
+		currentHeight := m.daIncludedHeight.Load()
+		if newHeight <= currentHeight {
+			break
+		}
+		if m.daIncludedHeight.CompareAndSwap(currentHeight, newHeight) {
+			heightBytes := make([]byte, 8)
+			binary.BigEndian.PutUint64(heightBytes, newHeight)
+			return m.store.SetMetadata(ctx, DAIncludedHeightKey, heightBytes)
+		}
+	}
+	return nil
+}
+
+// GetDAIncludedHeight returns the rollup height at which all blocks have been
+// included in the DA
+func (m *Manager) GetDAIncludedHeight() uint64 {
+	return m.daIncludedHeight.Load()
+}
